requirements.txt
# Core dependencies
smolagents[toolkit]>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.7.0
